refactor(heading): refactor package.json into index.js
rename heading.js to heading.component.js and correct appropriate imports
