+++ src/components/heading/heading.spec.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { shallow, mount } from "enzyme";
-import Heading from "./heading";
+import Heading from "./heading.component";
 import {
   StyledSubHeader,
   StyledDivider,
