+++ src/components/heading/heading.stories.mdx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
-import Heading from '.';
+import Heading from './heading.component';
 import Tile from '../tile';
 import Dl from '../definition-list/dl.component';
 import Dt from '../definition-list/dt.component';
