--- src/components/heading/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "./heading.js"
-}
