+++ src/components/pages/pages.stories.mdx
@@ -3,7 +3,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 import DefaultPages from './pages.component';
 import Page from './page/page.component';
 import DialogFullScreen from '../dialog-full-screen';
-import Heading from '../heading/heading';
+import Heading from '../heading';
 import Button from '../button';
 
 <Meta title="Pages" parameters={{ info: { disable: true }, chromatic: { disable: true }}} />
